restructure file structure to include types
