+++ src/components/pages/contact.js
@@ -3,7 +3,7 @@ import p5 from 'p5'
 import Matter from 'matter-js'
 import { connect } from 'react-redux'
 
-import { setupFrame, resetPageFrame, setupContact, createMouseConstraint } from '../utilities'
+import { setupFrame, resetPageFrame, setupContact, createMouseConstraint } from '../types'
 
 const { Engine, World } = Matter
