+++ src/components/types/index.js
@@ -1,3 +1,4 @@
 export * from './constructors'
 export * from './setups'
 export * from './utils'
+export * from './types'
