+++ src/components/types/utils.ts
@@ -2,22 +2,10 @@ import p5 from 'p5'
 import Matter from 'matter-js'
 
 import { 
-  Environment, 
   Color, 
   Position, 
-  Button, 
-  Body,
-  EngineType, 
-  WorldType,
-  Canvas,
-  Instance,
-  ColorRenderConfig,
-  ImageConfig,
-  TextConfig,
-  ParagraphConfig,
-  Constraint,
   TextSettings
-} from '../../types'
+} from './types'
 
 // import { ColorBallConstructor, BubbleConstructor } from './constructors'
 const { World, Mouse, MouseConstraint } = Matter
@@ -103,15 +91,6 @@ export const randomColor = (): Color => {
   }
 }
 
-export const desaturateColor = (environment: Environment, color: Color): Color => {
-  const { sketch } = environment
-  return {
-    hue: color.hue,
-    saturation: sketch.lerp(color.saturation, defaultColor.saturation, 0.4),
-    lightness: sketch.lerp(color.lightness, defaultColor.lightness, 0.2)
-  }
-}
-
 const areaFromPoints = (position: Position, vertices: Position[], sketch: p5): number => {
   // find and sum the triangles created from position and vertices
   return vertices
@@ -143,7 +122,7 @@ const areaFromPoints = (position: Position, vertices: Position[], sketch: p5): n
     }, 0)
 }
 
-export const createMouseConstraint = (canvas: Canvas, engine: EngineType, world: WorldType, sketch: p5) => {
+export const createMouseConstraint = (canvas: HTMLCanvasElement, engine: Matter.Engine, world: Matter.World, sketch: p5) => {
   const mouse = Mouse.create(canvas)
   mouse.pixelRatio = sketch.pixelDensity()
   const mouseOptions = {
@@ -151,6 +130,44 @@ export const createMouseConstraint = (canvas: Canvas, engine: EngineType, world:
   }
   const mouseConstraint = MouseConstraint.create(engine, mouseOptions)
   World.add(world, mouseConstraint)
+}
+
+export const transformBody = (sketch: p5, body: any) => {
+  const position = body.position
+  const angle = body.angle
+
+  sketch.translate(position.x, position.y)
+  sketch.rotate(angle)
+}
+
+export const setTextDimensions = (sketch: p5, textSettings: TextSettings) => {
+  // const { sketch, textSettings } = config
+  const { text, textSize, boxWidth, boxHeight, padding } = textSettings
+  sketch.textSize(textSize || 18)
+  return {
+    textSize: textSize,
+    w: boxWidth || sketch.textWidth(text),
+    h: boxHeight || sketch.textAscent(),
+    padding: padding || 10
+  }
+}
+
+export const renderText = (config: any) => {
+  const { sketch, textSettings, color, alignment } = config
+  const { textSize, text, boxWidth, boxHeight } = textSettings
+  const { hue, saturation, lightness } = color
+
+  sketch.rectMode('center')
+  sketch.textAlign(alignment.horizontal, alignment.vertical)
+  sketch.textSize(textSize)
+  sketch.colorMode('hsl')
+  sketch.fill(hue, saturation, lightness)
+  if (boxWidth && boxHeight) {
+    sketch.text(text, 0, 0, boxWidth, boxHeight)
+  }
+  else {
+    sketch.text(text, 0, 0)
+  }
 }
 
 // export const createColorParticles = (environment: Environment) => {
@@ -168,6 +185,15 @@ export const createMouseConstraint = (canvas: Canvas, engine: EngineType, world:
 //   }
 
 //   new ColorBallConstructor(environment, particleSettings)
+// }
+
+// export const desaturateColor = (environment: Environment, color: Color): Color => {
+//   const { sketch } = environment
+//   return {
+//     hue: color.hue,
+//     saturation: sketch.lerp(color.saturation, defaultColor.saturation, 0.4),
+//     lightness: sketch.lerp(color.lightness, defaultColor.lightness, 0.2)
+//   }
 // }
 
 // export const createBubbles = (environment: Environment, button: Button) => {
@@ -195,145 +221,107 @@ export const createMouseConstraint = (canvas: Canvas, engine: EngineType, world:
 //   })
 // }
 
-export const transformBody = (sketch: p5, body: any) => {
-  const position = body.position
-  const angle = body.angle
-
-  sketch.translate(position.x, position.y)
-  sketch.rotate(angle)
-}
-
-export const setTextDimensions = (sketch: p5, textSettings: TextSettings) => {
-  // const { sketch, textSettings } = config
-  const { text, textSize, boxWidth, boxHeight, padding } = textSettings
-  sketch.textSize(textSize || 18)
-  return {
-    textSize: textSize,
-    w: boxWidth || sketch.textWidth(text),
-    h: boxHeight || sketch.textAscent(),
-    padding: padding || 10
-  }
-}
-
 // export const addToWorld = (world: WorldType, instance: Instance, container: any[]) => {
 //   World.add(world, instance.body)
 //   container.push(instance)
 //   instance.index = container.length - 1
 // }
 
-export const renderText = (config: TextConfig | ParagraphConfig) => {
-  const { sketch, textSettings, color, alignment } = config
-  const { textSize, text, boxWidth, boxHeight } = textSettings
-  const { hue, saturation, lightness } = color
+// export const renderImage = (config: ImageConfig) => {
+//   const { sketch, image, dimensions } = config
 
-  sketch.rectMode('center')
-  sketch.textAlign(alignment.horizontal, alignment.vertical)
-  sketch.textSize(textSize)
-  sketch.colorMode('hsl')
-  sketch.fill(hue, saturation, lightness)
-  if (boxWidth && boxHeight) {
-    sketch.text(text, 0, 0, boxWidth, boxHeight)
-  }
-  else {
-    sketch.text(text, 0, 0)
-  }
-}
+//   sketch.imageMode('center')
+//   sketch.image(image, 0, 0, dimensions.w, dimensions.h)
+// }
 
-export const renderImage = (config: ImageConfig) => {
-  const { sketch, image, dimensions } = config
+// export const renderOutline = (config: ColorRenderConfig) => {
+//   const { sketch, color, dimensions, shape } = config
+//   const { hue, saturation, lightness } = color
+//   sketch.colorMode('hsl')
+//   sketch.noFill()
+//   sketch.stroke(hue, saturation, lightness)
+//   switch (shape) {
+//     case 'rect':
+//       sketch.rect(0, 0, dimensions.w + dimensions.padding, dimensions.h + dimensions.padding)
+//       break
+//     case 'circle':
+//       sketch.ellipse(0, 0, dimensions.w + dimensions.padding)
+//       break
+//   }
+// }
 
-  sketch.imageMode('center')
-  sketch.image(image, 0, 0, dimensions.w, dimensions.h)
-}
+// export const renderHighlight = (config: ColorRenderConfig) => {
+//   const { sketch, dimensions, shape } = config
+//   sketch.colorMode('hsl')
+//   sketch.noStroke()
+//   sketch.fill(0, 0, 100, 0.1)
+//   switch (shape) {
+//     case 'rect':
+//       sketch.rectMode('center')
+//       sketch.rect(0, 0, dimensions.w + (dimensions.padding || 0), dimensions.h + (dimensions.padding || 0))
+//       break
+//     case 'circle':
+//       sketch.ellipse(0, 0, dimensions.w + dimensions.padding)
+//       break
+//   }
+// }
 
-export const renderOutline = (config: ColorRenderConfig) => {
-  const { sketch, color, dimensions, shape } = config
-  const { hue, saturation, lightness } = color
-  sketch.colorMode('hsl')
-  sketch.noFill()
-  sketch.stroke(hue, saturation, lightness)
-  switch (shape) {
-    case 'rect':
-      sketch.rect(0, 0, dimensions.w + dimensions.padding, dimensions.h + dimensions.padding)
-      break
-    case 'circle':
-      sketch.ellipse(0, 0, dimensions.w + dimensions.padding)
-      break
-  }
-}
+// export const renderLowlight = (config: ColorRenderConfig) => {
+//   const { sketch, dimensions, shape } = config
+//   sketch.colorMode('hsl')
+//   sketch.noStroke()
+//   sketch.fill(0, 0, 0, 0.8)
+//   switch (shape) {
+//     case 'rect':
+//       sketch.rectMode('center')
+//       sketch.rect(0, 0, dimensions.w + (dimensions.padding || 0), dimensions.h + (dimensions.padding || 0))
+//       break
+//     case 'circle':
+//       sketch.ellipse(0, 0, dimensions.w + dimensions.padding)
+//       break
+//   }
+// }
 
-export const renderHighlight = (config: ColorRenderConfig) => {
-  const { sketch, dimensions, shape } = config
-  sketch.colorMode('hsl')
-  sketch.noStroke()
-  sketch.fill(0, 0, 100, 0.1)
-  switch (shape) {
-    case 'rect':
-      sketch.rectMode('center')
-      sketch.rect(0, 0, dimensions.w + (dimensions.padding || 0), dimensions.h + (dimensions.padding || 0))
-      break
-    case 'circle':
-      sketch.ellipse(0, 0, dimensions.w + dimensions.padding)
-      break
-  }
-}
-
-export const renderLowlight = (config: ColorRenderConfig) => {
-  const { sketch, dimensions, shape } = config
-  sketch.colorMode('hsl')
-  sketch.noStroke()
-  sketch.fill(0, 0, 0, 0.8)
-  switch (shape) {
-    case 'rect':
-      sketch.rectMode('center')
-      sketch.rect(0, 0, dimensions.w + (dimensions.padding || 0), dimensions.h + (dimensions.padding || 0))
-      break
-    case 'circle':
-      sketch.ellipse(0, 0, dimensions.w + dimensions.padding)
-      break
-  }
-}
-
-export const checkMouseInBounds = (body: Matter.Body, mousePosition: Position, config: ColorRenderConfig) => {
-  const { sketch, shape, dimensions } = config
-  // const { body } = instance
+// export const checkMouseInBounds = (body: Matter.Body, mousePosition: Position, config: ColorRenderConfig) => {
+//   const { sketch, shape, dimensions } = config
+//   // const { body } = instance
   
-  switch (shape) {
-    case 'rect':
-      const vertices = body.vertices
-      const mouseArea = areaFromPoints(mousePosition, vertices, sketch)
-      return (mouseArea < body.area + 1)
-    case 'circle':
-      const distance = sketch.dist(body.position.x, body.position.y, mousePosition.x, mousePosition.y)
-      return (distance < dimensions.w / 2)
-  }
-}
+//   switch (shape) {
+//     case 'rect':
+//       const vertices = body.vertices
+//       const mouseArea = areaFromPoints(mousePosition, vertices, sketch)
+//       return (mouseArea < body.area + 1)
+//     case 'circle':
+//       const distance = sketch.dist(body.position.x, body.position.y, mousePosition.x, mousePosition.y)
+//       return (distance < dimensions.w / 2)
+//   }
+// }
 
-export const manageParticleRender = (array: any[]) => {
-  for (let i = 0; i < array.length; i++) {
-    let particle = array[i]
-    particle.show()
-    if (particle.isBelowLine()) {
-      particle.remove()
-      array.splice(i, 1)
-      i--
-    }
-  }
-}
+// export const manageParticleRender = (array: any[]) => {
+//   for (let i = 0; i < array.length; i++) {
+//     let particle = array[i]
+//     particle.show()
+//     if (particle.isBelowLine()) {
+//       particle.remove()
+//       array.splice(i, 1)
+//       i--
+//     }
+//   }
+// }
 
-export const manageBubbleRender = (array: any[], mousePosition: Position) => {
-  for (let i = 0; i < array.length; i++) {
-    let bubble = array[i]
-    bubble.show()
-    bubble.checkMouseInBounds(mousePosition)
-    bubble.checkBubblePop()
-    if (bubble.bubbleShouldPop) {
-      bubble.remove()
-      array.splice(i, 1)
-      i--
-    }
-  }
-}
+// export const manageBubbleRender = (array: any[], mousePosition: Position) => {
+//   for (let i = 0; i < array.length; i++) {
+//     let bubble = array[i]
+//     bubble.show()
+//     bubble.checkMouseInBounds(mousePosition)
+//     bubble.checkBubblePop()
+//     if (bubble.bubbleShouldPop) {
+//       bubble.remove()
+//       array.splice(i, 1)
+//       i--
+//     }
+//   }
+// }
